bootloader: do not uninstall initrd if not previously installed
It fixes traceback when unapplying bootloader plugin tunig and there
is no initrd specified.

+++ tuned/plugins/plugin_bootloader.py
@@ -93,8 +93,9 @@ class BootloaderPlugin(base.Plugin):
 	def _remove_grub2_tuning(self):
 		self._patch_bootcmdline({consts.BOOT_CMDLINE_TUNED_VAR : "", consts.BOOT_CMDLINE_INITRD_ADD_VAR : ""})
 		self._cmd.add_modify_option_in_file(self._grub2_cfg_file_name, {"set\s+" + consts.GRUB2_TUNED_VAR : "", "set\s+" + consts.GRUB2_TUNED_INITRD_VAR : ""}, add = False)
-		log.info("removing initrd image '%s'" % self._initrd_dst_img_val)
-		self._cmd.unlink(self._initrd_dst_img_val)
+		if self._initrd_dst_img_val is not None:
+			log.info("removing initrd image '%s'" % self._initrd_dst_img_val)
+			self._cmd.unlink(self._initrd_dst_img_val)
 
 	def _instance_unapply_static(self, instance, profile_switch = False):
 		if profile_switch:
